Windows build support with CMake

Approved-by: Baptiste MOUGINOT (BaM)
Approved-by: Iulian Grindeanu

+++ src/moab/Util.hpp
@@ -63,7 +63,7 @@ template <typename T>
 inline
 bool Util::is_finite(T value)
 {
-  return moab_isfinite(value);
+  return moab_isfinite((double)value);
 }
 
 } // namespace moab
